Revert "[TID-170]Feature/blockchain foundry clean (#143)" (#145)
This reverts commit 4f4ccffbda65e8347e68a2a0a2850dce0fd8b465.
