+++ .env.example
@@ -29,8 +29,8 @@ DATABASE_URL="postgres://db_user:db_password@localhost:5432/transcendence"
 DEBUG=True
 NODE_ENV=development
 
-FOUNDRY_HOST=127.0.0.1
-FOUNDRY_PORT=8545
-FOUNDRY_URL=http://127.0.0.1:8545/
-FOUNDRY_MNEMONIC="test test test test test test test test test test test junk"
-FOUNDRY_PRIVATE_KEY=0x59c6995e998f97a5a0044966f0945388cf5f3d99b02aa0055b1577e5c9b8c7fa
+GANACHE_HOST=127.0.0.1
+GANACHE_PORT=8545    
+GANACHE_NETWORK_ID=5777
+GANACHE_URL=http://127.0.0.1:8545
+GANACHE_COD="bal bal bal bal bal"
