+++ .gitignore
@@ -92,18 +92,3 @@ schema.json
 #file blockchain
 blockchain/deployedAddress.json
 menu.sh
-blockchain/__pycache__/
-blockchain/*.pyc
-blockchain/anvil.log
-blockchain/*.log
-blockchain/cache/
-blockchain/broadcast/
-blockchain/out/
-blockchain/lib/forge-std/test/
-blockchain/lib/forge-std/CONTRIBUTING.md
-blockchain/lib/forge-std/README.md
-blockchain/lib/forge-std/cache/
-blockchain/lib/forge-std/out/
-blockchain/lib/forge-std.vscode
-blockchain/lib/forge-std.idea
-
