--- blockchain/foundry.toml
@@ -1,7 +1,0 @@
-[profile.default]
-src = "src"
-out = "out"
-libs = ["lib"]
-remappings = [
-  "forge-std/=lib/forge-std/src/"
-]
