--- blockchain/lib/forge-std/.gitattributes
@@ -1,1 +1,0 @@
-src/Vm.sol linguist-generated
