--- blockchain/lib/forge-std/.gitignore
@@ -1,4 +1,0 @@
-cache/
-out/
-.vscode
-.idea
