--- blockchain/lib/forge-std/foundry.toml
@@ -1,23 +1,0 @@
-[profile.default]
-fs_permissions = [{ access = "read-write", path = "./"}]
-optimizer = true
-optimizer_runs = 200
-
-[rpc_endpoints]
-# The RPC URLs are modified versions of the default for testing initialization.
-mainnet = "https://eth-mainnet.alchemyapi.io/v2/WV407BEiBmjNJfKo9Uo_55u0z0ITyCOX" # Different API key.
-optimism_sepolia = "https://sepolia.optimism.io/" # Adds a trailing slash.
-arbitrum_one_sepolia = "https://sepolia-rollup.arbitrum.io/rpc/" # Adds a trailing slash.
-needs_undefined_env_var = "${UNDEFINED_RPC_URL_PLACEHOLDER}"
-
-[fmt]
-# These are all the `forge fmt` defaults.
-line_length = 120
-tab_width = 4
-bracket_spacing = false
-int_types = 'long'
-multiline_func_header = 'attributes_first'
-quote_style = 'double'
-number_underscore = 'preserve'
-single_line_statement_blocks = 'preserve'
-ignore = ["src/console.sol", "src/console2.sol"]
