--- blockchain/lib/forge-std/src/Script.sol
@@ -1,27 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.6.2 <0.9.0;
-
-// 💬 ABOUT
-// Forge Std's default Script.
-
-// 🧩 MODULES
-import {console} from "./console.sol";
-import {console2} from "./console2.sol";
-import {safeconsole} from "./safeconsole.sol";
-import {StdChains} from "./StdChains.sol";
-import {StdCheatsSafe} from "./StdCheats.sol";
-import {stdJson} from "./StdJson.sol";
-import {stdMath} from "./StdMath.sol";
-import {StdStorage, stdStorageSafe} from "./StdStorage.sol";
-import {StdStyle} from "./StdStyle.sol";
-import {StdUtils} from "./StdUtils.sol";
-import {VmSafe} from "./Vm.sol";
-
-// 📦 BOILERPLATE
-import {ScriptBase} from "./Base.sol";
-
-// ⭐️ SCRIPT
-abstract contract Script is ScriptBase, StdChains, StdCheatsSafe, StdUtils {
-    // Note: IS_SCRIPT() must return true.
-    bool public IS_SCRIPT = true;
-}
