--- blockchain/lib/forge-std/src/StdError.sol
@@ -1,15 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Panics work for versions >=0.8.0, but we lowered the pragma to make this compatible with Test
-pragma solidity >=0.6.2 <0.9.0;
-
-library stdError {
-    bytes public constant assertionError = abi.encodeWithSignature("Panic(uint256)", 0x01);
-    bytes public constant arithmeticError = abi.encodeWithSignature("Panic(uint256)", 0x11);
-    bytes public constant divisionError = abi.encodeWithSignature("Panic(uint256)", 0x12);
-    bytes public constant enumConversionError = abi.encodeWithSignature("Panic(uint256)", 0x21);
-    bytes public constant encodeStorageError = abi.encodeWithSignature("Panic(uint256)", 0x22);
-    bytes public constant popError = abi.encodeWithSignature("Panic(uint256)", 0x31);
-    bytes public constant indexOOBError = abi.encodeWithSignature("Panic(uint256)", 0x32);
-    bytes public constant memOverflowError = abi.encodeWithSignature("Panic(uint256)", 0x41);
-    bytes public constant zeroVarError = abi.encodeWithSignature("Panic(uint256)", 0x51);
-}
