--- blockchain/lib/forge-std/src/console2.sol
@@ -1,4 +1,0 @@
-// SPDX-License-Identifier: MIT
-pragma solidity >=0.4.22 <0.9.0;
-
-import {console as console2} from "./console.sol";
