--- hashiCorpVault/data/.root-token
@@ -1,1 +1,0 @@
-hvs.vTZWPd79zCk7LwIbkLpSGgVT
