--- reverse_proxy/certs/localhost.crt
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDrzCCApegAwIBAgIUZer5b+ekYQNKnfL2VXSXiwSmQlswDQYJKoZIhvcNAQEL
-BQAwFDESMBAGA1UEAxMJbG9jYWxob3N0MB4XDTI1MDQxMzE3MTA1NFoXDTI1MDQx
-NjE3MTEyNFowFDESMBAGA1UEAxMJbG9jYWxob3N0MIIBIjANBgkqhkiG9w0BAQEF
-AAOCAQ8AMIIBCgKCAQEA3buQxHpcuwzbHaeAp3M5/sgvoCKXmWM4BbF6INoCi6jX
-zC1vez3ts1NtrKRD6wyPtBioZf75VBBQEph3axrPWQnCxMV9LfOFHUQ5qsULzx6c
-YlMS6mkdLRY+XpnuYne/dJHgBXmGAoKl9rPKMsEcAh9tr8z8GGlHnDwIkC7CB/zt
-bBLkvfdBI85QMRGghdj2gYu+W2AibR/O6Z5G0j9vrx2/dUxpdxqMmxeNyigAT6XK
-P24vGeGB1RGmlnHNp8hVGSgCWfeH31Qy3VIK+wXQvEKWvmLUFGRnLiTZkracwgIp
-guJeFQQIJX7Fpz2nJbVL4gk/BVrBtIu1jb8lVB05gQIDAQABo4H4MIH1MA4GA1Ud
-DwEB/wQEAwIDqDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwHQYDVR0O
-BBYEFHuYaO2P4BnH7Er9tbVYB5FlVzk3MB8GA1UdIwQYMBaAFCTRGKq+8f76pEf3
-SJfTovoT9DULMDsGCCsGAQUFBwEBBC8wLTArBggrBgEFBQcwAoYfaHR0cDovLzEy
-Ny4wLjAuMTo4MjAwL3YxL3BraS9jYTAUBgNVHREEDTALgglsb2NhbGhvc3QwMQYD
-VR0fBCowKDAmoCSgIoYgaHR0cDovLzEyNy4wLjAuMTo4MjAwL3YxL3BraS9jcmww
-DQYJKoZIhvcNAQELBQADggEBACERoErz8BX3NDzaO3ad1sfNr5AmnXH8N0QHzE7E
-fSy+VqnYIxySNKGy5l0JC2ODs0lvaXY+NBSYiH8bWBK5WH4pnMK/cNqxfYoSWO6j
-DKsdQUoqQkmLrdauMUy39cQlDt+ggm069me4klNljU0iSdiavL4R0YWRf5FkCJFM
-WzDMg3zIYEEtMXuYh309Z8Fjr/NCHlVzH8JpqkcMhKZ8sVAGbyDnXnBehN/E88SG
-GLTBC4KnSgBtajxHfT6Zc4nraxDApCupOwnhcGt6bqN58Mt1b3HaAXP6is967JMU
-VZ0iS1u6VoTMHYDOu9lmYl/AzIOpthiX+mWr9zseCNp02g0=
------END CERTIFICATE-----
