--- reverse_proxy/certs/localhost.key
@@ -1,27 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIEpQIBAAKCAQEA3buQxHpcuwzbHaeAp3M5/sgvoCKXmWM4BbF6INoCi6jXzC1v
-ez3ts1NtrKRD6wyPtBioZf75VBBQEph3axrPWQnCxMV9LfOFHUQ5qsULzx6cYlMS
-6mkdLRY+XpnuYne/dJHgBXmGAoKl9rPKMsEcAh9tr8z8GGlHnDwIkC7CB/ztbBLk
-vfdBI85QMRGghdj2gYu+W2AibR/O6Z5G0j9vrx2/dUxpdxqMmxeNyigAT6XKP24v
-GeGB1RGmlnHNp8hVGSgCWfeH31Qy3VIK+wXQvEKWvmLUFGRnLiTZkracwgIpguJe
-FQQIJX7Fpz2nJbVL4gk/BVrBtIu1jb8lVB05gQIDAQABAoIBAQCzz1kUnA36+e/u
-2ljwcfbD8F7+w06g6t8Rbt0VKvYJf4CQaQAQtyW92JgF+qfBQsx+Rci5/cz/KMNm
-cRHWX0VqdQnnPCfJTcPS+fKx011CNxamfavHQDqSx4kIelfgri8qPummRugvvPcN
-WrtL8WHfLtFqURbWTPnvbpebJO/gPO9rStIECABw81DDvRAnKjhMk0ft/ZceBTjg
-6s/82QWo12J+ebyMBFiV1SRa8WjAyytgsV6ErsJdiNpLZKq9EtCoPpCk9A4Q5xPf
-e1f1LH8S0umDOrwyKVWujMVO+baUjn9LMNB1Z4kP8ju9RJkItPIlcIGY2e+1fO3/
-x9F6RjRpAoGBAPJZ8hYucdgWOEOLoEGrjAEnV9nhl+MdA96+7B86bfbwqiSR/WdL
-mQFepUFh0Wq+jzf9Mt7nZ4rr4AePmnOHFeVCd49B2vFji2gIu15zHPM6XbSS4HIy
-U6rAAfd4ypt/LXIxu01uW6CROUK2pyy2utldByBbSwJJjJPgqQVLRQkfAoGBAOo4
-WqRjo4/fc2rIdnK0/Km2l966UlVbZfrA8l6Vqj/2d6Ka9RdHZEX8usXe38WF5N6Z
-EUjdliXnQt3ZCY2pDnhg9TYuhrgLEBud+mWB+43AiwH9ZosXobxYFQy4+zWSTg4U
-syIT24APrzSyN3o+grRJtXfMt3hcjWqLBaUse0lfAoGAfjuLdTJL6hnoxIX9uzb3
-fcJYaAyxaIK8eVFI/BsbG/zPIq1IKH2f/K6ntR6WTdtkDe28iY5OyHWEA2wxLuu1
-4+xzChNrT0BiZJXnwpKDCc3k8SFv9wtvuQi5/YTVQ5lCurBct0A8Sa5a/PzWptun
-B4EFLT1gHJjiuEPUxgUmuFECgYEAuqNDX2FikZ/uebXZ5Rfhil9M75erSgc14TXl
-LIe4jZWZyYIPMYibF8YHEuvBhsVmFhcFFfo0zmlU+DEdkruXI0envpAenAN8x3UA
-dei5CkDWxDQnhUjEga/yvhudWhgGiOPnxozqSUa2R5iH6qprPrzJlPgVbxzZN4Am
-dHNYnqMCgYEAmdnbmlFjpQKIiQQfa52xVvoIgnCKmOZMgSGZDryY5OUQjni2oEp7
-ocBYafne6l2Ok028nxqgW8jWCVF5aScoNxf4+SF5wu3FqgvB5i3hW7WQwmYN8YDa
-ZU2NWaaWpfFZvyREUUVa8QHbdVag9Ese88Xm9Qyr0j+rJy3mzvKykX0=
------END RSA PRIVATE KEY-----
